@@ -5,11 +5,11 @@ import org.telegram.telegrambots.meta.api.objects.Chat;
 import org.telegram.telegrambots.meta.api.objects.User;
 import org.telegram.telegrambots.meta.bots.AbsSender;
 
-public class StatusCmd extends BotCommand {
+public class StatusCommand extends BotCommand {
     private static final String COMMAND_IDENTIFIER = "status";
     private static final String COMMAND_DESCRIPTION = "shows user status";
 
-    public StatusCmd(){
+    public StatusCommand(){
         super(COMMAND_IDENTIFIER, COMMAND_DESCRIPTION);
     }
 
