@@ -5,11 +5,11 @@ import org.telegram.telegrambots.meta.api.objects.Chat;
 import org.telegram.telegrambots.meta.api.objects.User;
 import org.telegram.telegrambots.meta.bots.AbsSender;
 
-public class TutorialCmd extends BotCommand {
+public class TutorialCommand extends BotCommand {
     private static final String COMMAND_IDENTIFIER = "tutorial";
     private static final String COMMAND_DESCRIPTION = "learn to use this bot";
 
-    public TutorialCmd(String commandIdentifier, String description) {
+    public TutorialCommand(String commandIdentifier, String description) {
         super(COMMAND_IDENTIFIER, COMMAND_DESCRIPTION);
     }
 
