@@ -1,10 +1,12 @@
 package com.alexeykovzel.bot.feature.viewlist;
 
-import com.alexeykovzel.bot.feature.query.QueryType;
-import com.alexeykovzel.bot.feature.query.BasicQuery;
+import com.alexeykovzel.bot.query.*;
 import com.alexeykovzel.db.service.CaseStudyDataService;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import org.telegram.telegrambots.meta.api.methods.AnswerCallbackQuery;
+import org.telegram.telegrambots.meta.api.methods.updatingmessages.EditMessageReplyMarkup;
+import org.telegram.telegrambots.meta.api.methods.updatingmessages.EditMessageText;
 import org.telegram.telegrambots.meta.api.objects.CallbackQuery;
 import org.telegram.telegrambots.meta.api.objects.Message;
 import org.telegram.telegrambots.meta.bots.AbsSender;
@@ -13,69 +15,76 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-public class ViewListQuery extends BasicQuery {
-    private static final QueryType type = QueryType.VIEW_LIST;
-    public final static int maxTermsPerPage = 5;
-
+public class ViewListQuery extends QueryTemplate {
     private final CaseStudyDataService caseStudyDataService;
-    private ViewListStatus status;
+    public final int termsPerPage = ViewListBuilder.defaultTermsPerPage;
 
-    public ViewListQuery(String[] args, AbsSender absSender, CallbackQuery basicQuery,
-                         CaseStudyDataService caseStudyDataService) {
-        super(args, absSender, basicQuery);
+    public ViewListQuery(CaseStudyDataService caseStudyDataService) {
         this.caseStudyDataService = caseStudyDataService;
     }
 
     @Override
-    public void execute() {
+    public void execute(AbsSender absSender, QueryDto queryDto, CallbackQuery basicQuery) {
         Message message = basicQuery.getMessage();
         String chatId = message.getChatId().toString();
         Integer messageId = message.getMessageId();
+        String[] args = queryDto.getArgs();
+        ViewListStatus status = ViewListStatus.fromKey(queryDto.getStatusKey());
 
-        switch (status) {
-            case DEFAULT:
-                int page = Integer.parseInt(args[0]);
-                Optional<List<String>> optTermValues = caseStudyDataService.findTermValuesByChatId(chatId);
-                optTermValues.ifPresent(terms -> editMessageReplyMarkup(chatId,
-                        messageId, ViewListBuilder.getListViewMarkup(terms, page)));
-                break;
-            case PAGE_PANEL:
-                optTermValues = caseStudyDataService.findTermValuesByChatId(chatId);
-                optTermValues.ifPresent(values -> {
-                    int pagesSum = values.size() / maxTermsPerPage;
-                    editMessage(chatId, messageId,
-                            "Please select a page",
-                            ViewListBuilder.getPagePanelMarkup(pagesSum, 7));
-                });
-                break;
-            case ABC_VIEW:
-                page = Integer.parseInt(args[0]);
-                char letter = args[1].charAt(0);
-                optTermValues = caseStudyDataService.findTermValuesByChatId(chatId);
-                optTermValues.ifPresent(terms -> {
-                    List<String> termsByLetter = terms.stream()
-                            .collect(Collectors.partitioningBy(s -> s.charAt(0) == letter)).get(Boolean.TRUE);
-                    editMessageReplyMarkup(chatId, messageId, ViewListBuilder.getListViewMarkup(termsByLetter, page));
-                });
-                break;
-            case ABC_PANEL:
-                editMessageReplyMarkup(chatId, messageId, ViewListBuilder.getAlphabeticalMarkup(7));
-                break;
+        if (status != null) {
+            switch (status) {
+                case DEFAULT:
+                    int page = Integer.parseInt(args[0]);
+                    Optional<List<String>> optTermValues = caseStudyDataService.findTermValuesByChatId(chatId);
+                    optTermValues.ifPresent(terms -> executeApiMethod(absSender, EditMessageReplyMarkup.builder()
+                            .chatId(chatId).messageId(messageId)
+                            .replyMarkup(ViewListBuilder.getListViewMarkup(terms, page, termsPerPage)).build()));
+                    break;
+                case PAGE_PANEL:
+                    optTermValues = caseStudyDataService.findTermValuesByChatId(chatId);
+                    optTermValues.ifPresent(values -> {
+                        int pagesSum = values.size() / termsPerPage;
+
+                        executeApiMethod(absSender, EditMessageText.builder().chatId(chatId).messageId(messageId)
+                                .text("Please select a page")
+                                .replyMarkup(ViewListBuilder.getPagePanelMarkup(pagesSum,
+                                        ViewListBuilder.defaultItemsPerRow)).build());
+                    });
+                    break;
+                case ABC_VIEW:
+                    page = Integer.parseInt(args[0]);
+                    char letter = args[1].charAt(0);
+                    optTermValues = caseStudyDataService.findTermValuesByChatId(chatId);
+                    optTermValues.ifPresent(terms -> {
+                        List<String> termsByLetter = terms.stream()
+                                .collect(Collectors.partitioningBy(s -> s.charAt(0) == letter)).get(Boolean.TRUE);
+
+                        executeApiMethod(absSender, EditMessageReplyMarkup.builder()
+                                .chatId(chatId).messageId(messageId)
+                                .replyMarkup(ViewListBuilder.getListViewMarkup(termsByLetter,
+                                        page, termsPerPage)).build());
+                    });
+                    break;
+                case ABC_PANEL:
+                    executeApiMethod(absSender, EditMessageReplyMarkup.builder()
+                            .chatId(chatId).messageId(messageId)
+                            .replyMarkup(ViewListBuilder.getAlphabeticalMarkup(7)).build());
+                    break;
+            }
+            AnswerCallbackQuery answerCallbackQuery = AnswerCallbackQuery.builder()
+                    .callbackQueryId(basicQuery.getId()).build();
+            executeApiMethod(absSender, answerCallbackQuery);
         }
-        sendAnswerCallbackQuery(basicQuery.getId()); // To remove loading icon
     }
 
     @Override
-    public void setStatusByKey(String key) {
-        ViewListStatus status = ViewListStatus.fromKey(key);
-        if (status != null) {
-            this.status = status;
-        }
+    public QueryType getType() {
+        return QueryType.VIEW_LIST;
     }
 
     @AllArgsConstructor
-    public enum ViewListStatus {
-        PAGE_PANEL("0"), DEFAULT("1"), ABC_VIEW("2"), ABC_PANEL("3");
+    enum ViewListStatus {
+        DEFAULT("0"), PAGE_PANEL("1"), ABC_VIEW("2"), ABC_PANEL("3");
 
         @Getter
         private String key;
