@@ -1,11 +1,10 @@
 package com.alexeykovzel.bot.handler;
 
-import com.alexeykovzel.bot.handler.abs.MessageHandler;
 import org.telegram.telegrambots.meta.api.objects.Message;
 import org.telegram.telegrambots.meta.api.objects.Update;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 
-public interface BotHandler extends MessageHandler {
+public interface BotHandler {
     void handleUpdate(Update update) throws TelegramApiException;
 
     void handleInvalidCommandUpdate(Update update);
