@@ -1,24 +1,29 @@
 package com.alexeykovzel.bot.handler.localhost;
 
+import com.alexeykovzel.bot.query.QueryRegistry;
 import com.alexeykovzel.bot.handler.BotHandler;
 import org.telegram.telegrambots.bots.DefaultBotOptions;
 import org.telegram.telegrambots.bots.TelegramLongPollingBot;
 import org.telegram.telegrambots.extensions.bots.commandbot.commands.CommandRegistry;
 import org.telegram.telegrambots.meta.api.methods.AnswerCallbackQuery;
+import org.telegram.telegrambots.meta.api.methods.BotApiMethod;
 import org.telegram.telegrambots.meta.api.methods.ParseMode;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.methods.updatingmessages.DeleteMessage;
 import org.telegram.telegrambots.meta.api.methods.updatingmessages.EditMessageReplyMarkup;
 import org.telegram.telegrambots.meta.api.methods.updatingmessages.EditMessageText;
-import org.telegram.telegrambots.meta.api.objects.CallbackQuery;
 import org.telegram.telegrambots.meta.api.objects.Message;
 import org.telegram.telegrambots.meta.api.objects.Update;
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.InlineKeyboardMarkup;
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.InlineKeyboardButton;
+import org.telegram.telegrambots.meta.bots.AbsSender;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
+
+import java.io.Serializable;
 
 abstract class LongPollingBotHandler extends TelegramLongPollingBot implements BotHandler {
     protected CommandRegistry commandRegistry;
+    protected QueryRegistry queryRegistry;
 
     public LongPollingBotHandler() {
         this(new DefaultBotOptions());
@@ -31,6 +36,7 @@ abstract class LongPollingBotHandler extends TelegramLongPollingBot implements B
     public LongPollingBotHandler(DefaultBotOptions options, boolean allowCommandsWithUsername) {
         super(options);
         this.commandRegistry = new CommandRegistry(allowCommandsWithUsername, this::getBotUsername);
+        this.queryRegistry = new QueryRegistry();
     }
 
     @Override
@@ -55,112 +61,11 @@ abstract class LongPollingBotHandler extends TelegramLongPollingBot implements B
         return false;
     }
 
-    @Override
-    public synchronized void sendMsg(String chatId, String text, InlineKeyboardMarkup inlineKeyboardMarkup) {
-        SendMessage sendMessage = SendMessage.builder()
-                .chatId(chatId)
-                .text(text)
-                .parseMode(ParseMode.MARKDOWN)
-                .disableWebPagePreview(true)
-                .replyMarkup(inlineKeyboardMarkup).build();
+    protected <T extends Serializable, Method extends BotApiMethod<T>> void executeApiMethod(Method method) {
         try {
-            execute(sendMessage);
+            execute(method);
         } catch (TelegramApiException e) {
             e.getStackTrace();
         }
-    }
-
-    @Override
-    public synchronized void sendMsg(String chatId, String text) {
-        SendMessage sendMessage = SendMessage.builder()
-                .chatId(chatId)
-                .text(text)
-                .parseMode(ParseMode.MARKDOWN)
-                .disableWebPagePreview(true).build();
-        try {
-            execute(sendMessage);
-        } catch (TelegramApiException e) {
-            e.getStackTrace();
-        }
-    }
-
-    @Override
-    public synchronized void deleteMsg(String chatId, Integer messageId) {
-        DeleteMessage deleteMessage = DeleteMessage.builder()
-                .chatId(chatId)
-                .messageId(messageId).build();
-        try {
-            execute(deleteMessage);
-        } catch (TelegramApiException e) {
-            e.getStackTrace();
-        }
-    }
-
-    @Override
-    public synchronized void sendAnswerCallbackQuery(String text, boolean alert, String callbackQueryId) {
-        AnswerCallbackQuery answerCallbackQuery = AnswerCallbackQuery.builder()
-                .callbackQueryId(callbackQueryId)
-                .showAlert(alert)
-                .text(text).build();
-        try {
-            execute(answerCallbackQuery);
-        } catch (TelegramApiException e) {
-            e.getStackTrace();
-        }
-    }
-
-    @Override
-    public synchronized void sendAnswerCallbackQuery(String callbackQueryId) {
-        AnswerCallbackQuery answerCallbackQuery = AnswerCallbackQuery.builder()
-                .callbackQueryId(callbackQueryId).build();
-        try {
-            execute(answerCallbackQuery);
-        } catch (TelegramApiException e) {
-            e.getStackTrace();
-        }
-    }
-
-    @Override
-    public void editMessageReplyMarkup(String chatId, Integer messageId, InlineKeyboardMarkup markup) {
-        EditMessageReplyMarkup replyMarkup = EditMessageReplyMarkup.builder()
-                .chatId(chatId)
-                .messageId(messageId)
-                .replyMarkup(markup).build();
-        try {
-            execute(replyMarkup);
-        } catch (TelegramApiException e) {
-            e.getStackTrace();
-        }
-    }
-
-    @Override
-    public void editMessage(String chatId, Integer messageId, String text, InlineKeyboardMarkup markup) {
-        EditMessageText replyMarkup = EditMessageText.builder()
-                .chatId(chatId)
-                .messageId(messageId)
-                .text(text)
-                .replyMarkup(markup).build();
-        try {
-            execute(replyMarkup);
-        } catch (TelegramApiException e) {
-            e.getStackTrace();
-        }
-    }
-
-    @Override
-    public String escapeMarkdown(String text) {
-        return text
-                .replace("_", "\\_")
-                .replace("*", "\\*")
-                .replace("[", "\\[")
-                .replace("`", "\\`");
-    }
-
-    @Override
-    public InlineKeyboardButton createInlineKeyboardButton(String text, String callbackData) {
-        InlineKeyboardButton inlineKeyboardButton = new InlineKeyboardButton();
-        inlineKeyboardButton.setText(text);
-        inlineKeyboardButton.setCallbackData(callbackData);
-        return inlineKeyboardButton;
     }
 }
